% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit_analysis.R
\name{digit_analysis}
\alias{digit_analysis}
\title{Digit analysis}
\usage{
digit_analysis(x, type = "terminal")
}
\arguments{
\item{x}{A vector or matrix of numeric values.}

\item{type}{Type of digit analysis ('benford' or 'terminal')}
}
\value{
P-value
}
\description{
Digit analysis
}
\examples{
 digit_analysis(c(1.234, 65.4321, 53.222), type = 'terminal')
 digit_analysis(c(1.234, 65.4321, 53.222), type = 'benford')
}
