% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_method.R
\name{fisher_method}
\alias{fisher_method}
\title{Fisher method}
\usage{
fisher_method(pval, threshold = 0, method = "default")
}
\arguments{
\item{pval}{Vector of pvalues}

\item{threshold}{Lowerbound threshold of p-values}

\item{method}{Method ('default' or 'reversed')}
}
\value{
Fisher method results
}
\description{
Fisher method
}
