% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit_analysis.R
\name{plot_digit_analysis}
\alias{plot_digit_analysis}
\title{Plot of digit analysis}
\usage{
plot_digit_analysis(x, type = "terminal")
}
\arguments{
\item{x}{A vector or matrix of numeric values.}

\item{type}{Type of digit analysis ('benford' or 'terminal')}
}
\value{
Boolean
}
\description{
Plot of digit analysis
}
\examples{
 plot_digit_analysis(c(1.234, 65.4321, 53.222), type = 'terminal')
 plot_digit_analysis(c(1.234, 65.4321, 53.222), type = 'benford')
}
