% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_var.R
\name{std_var}
\alias{std_var}
\title{Variance analysis}
\usage{
std_var(n, sds, subgroups, iter = 1000, method = "sd")
}
\arguments{
\item{n}{Vector of N per subgroup}

\item{sds}{Vector of SDs per subgroup}

\item{subgroups}{Factor vector of grouping}

\item{iter}{Number of iterations}

\item{method}{Variance analysis operationalization (opt 'sd' or 'maxmin')}
}
\value{
P-value
}
\description{
Variance analysis
}
\examples{
  std_var(n = rep(15, 3), sds = c(25.09, 24.58, 25.65), iter = 100, subgroups = rep(1, 3))
}
